# Optimize delta prefix search