This PR fixes a somewhat broken implementation of the binary search of prefix lengths in the DELTA_BYTE_ARRAY encoding.

```
name                                old time/op  new time/op  delta
BinarySearchPrefixLength/size=10    14.7ns ± 0%  11.3ns ± 2%  -23.21%  (p=0.000 n=9+10)
BinarySearchPrefixLength/size=100   41.1ns ± 2%  19.6ns ± 6%  -52.34%  (p=0.000 n=10+10)
BinarySearchPrefixLength/size=1000  88.3ns ± 1%  27.6ns ± 4%  -68.71%  (p=0.000 n=9+10)
```